+++ argopy/related/__init__.py
@@ -4,7 +4,7 @@ from .reference_tables import ArgoNVSReferenceTables
 from .argo_documentation import ArgoDocs
 from .doi_snapshot import ArgoDOI
 from .euroargo_api import get_coriolis_profile_id, get_ea_profile_page
-
+from .utils import load_dict, mapp_dict
 
 #
 __all__ = (
@@ -18,4 +18,8 @@ __all__ = (
     # Functions:
     "get_coriolis_profile_id",
     "get_ea_profile_page",
+
+    # Utilities:
+    "load_dict",
+    "mapp_dict",
 )
