+++ argopy/stores/argo_index_proto.py
@@ -505,7 +505,7 @@ class ArgoIndexStoreProto(ABC):
         else:
             log.debug("Converting [%s] to dataframe from scratch ..." % src)
             # Post-processing for user:
-            from argopy.utilities import load_dict, mapp_dict
+            from ..related import load_dict, mapp_dict
 
             if nrows is not None:
                 df = df.loc[0: nrows - 1].copy()
